Implementation for AllocateEid responder
This implementation is for support the decode/responder
for allocate EID command.

Bus owners are responsible for allocating pool of EID to MCTP bridge
The bus owner shall use Allocate Eid command to assign the pool
size to the bridge. Response to the SetEndpointId indicates whether
the endpoint supports an EID pool.

Signed-off-by: deepak kumar agrawal <deepakx.agrawal@intel.com>