smbus_bridge: Unregister devices that no longer available (#92)
Currently in SMBusBridge::initEndpointDiscovery(), it only checks if
there are new SMBus devices and try to register them. Device removal
is not handled. This commit fixes it by checking if a registered device
still exists on each scan and unregister it when it is no longer found
on system.

Tested:
On system with two U.2 SSDs with MCTP support installed, unplug one,
verified the unplugged SSD is unregistered on next scan.

Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>