Cleanup file closure procedure
Boost asio documentation captures that posix_descriptor::release() does
not close the underlying file descriptor and the it has to be manually
closed. This patch handles the same. There is no difference in feature behaviour
before and after the change.

Tested:
Power cycled the system and ensured that I3C interface gets
re-initialised.

Signed-off-by: Sumanth Bhat <sumanth.bhat@intel.com>