Inotify, Avoid unwanted bus rescan (#157)
During an inotify event mctpd rescans and refresh devices on i2c ports.
This results in re-registering of all devices even if it is a
pre-existing
device with no change. This is happening because Fd to existing device
also
gets updated as part of re-scan.

Ideally speaking we don't need to trigger a rescan for all inotify
event.
We just need to rescan only if the new addition/deletion of i2c device
has
the same rootbus the daemon is serving. This change adds the filtering
logic to avoid unwanted bus rescans.

Added more delay before mctp re-discovery so that other services will
get
first chance to complete i2c scan. Which helps to avoid i2c traffic
congestion
in a complex i2c bus architecture.

Tested: Verified mctpd re-discovery is skipped when an inotify event
from
an independent i2c device(not under current daemon preview) is received.

Signed-off-by: Arun P. Mohanan <arun.p.m@intel.com>