Fix: Multiple getDriverFD() calls in parallel (#400)
When PFR device is not ready for ENTDAA getDriverFD() waits for 20
seconds for PFR to come up. In this time if espi reset is received then
two rescan loops will come in parallel. Increasing the timeout to 60
seconds so that second coroutine will wait enough.

Also HPM CPU service names are added to AllowList but it is not sure all
4 will be present. Fix the same in GetUniquename function

Signed-off-by: Nidhin MS <nidhin.ms@intel.com>